Fixes #727: Migrate from Javassist to ByteBuddy: Refactoring MockTransformer  - split transformer to set of small classes.
