+++ powermock-core/src/main/java/org/powermock/core/classloader/MockClassLoader.java
@@ -20,6 +20,8 @@ import org.powermock.core.transformers.ClassWrapper;
 import org.powermock.core.transformers.ClassWrapperFactory;
 import org.powermock.core.transformers.MockTransformer;
 import org.powermock.tests.utils.IgnorePackagesExtractor;
+import org.powermock.core.transformers.MockTransformerChain;
+import org.powermock.core.transformers.javassist.support.ClassWrapperFactoryImpl;
 
 import java.security.ProtectionDomain;
 import java.util.List;
@@ -55,7 +57,7 @@ public abstract class MockClassLoader extends DeferSupportingClassLoader {
     
     protected ClassMarker classMarker;
     protected ClassWrapperFactory classWrapperFactory;
-    private List<MockTransformer> mockTransformerChain;
+    private MockTransformerChain mockTransformerChain;
     
     /**
      * Creates a new instance of the  based on the
@@ -79,7 +81,7 @@ public abstract class MockClassLoader extends DeferSupportingClassLoader {
      */
     protected MockClassLoader(MockClassLoaderConfiguration configuration) {
         super(MockClassLoader.class.getClassLoader(), configuration);
-        classWrapperFactory = new ClassWrapperFactory();
+        classWrapperFactory = new ClassWrapperFactoryImpl();
     }
     
     @Override
@@ -95,7 +97,7 @@ public abstract class MockClassLoader extends DeferSupportingClassLoader {
         return loadedClass;
     }
     
-    public void setMockTransformerChain(List<MockTransformer> mockTransformerChain) {
+    public void setMockTransformerChain(MockTransformerChain mockTransformerChain) {
         this.mockTransformerChain = mockTransformerChain;
     }
     
@@ -105,9 +107,8 @@ public abstract class MockClassLoader extends DeferSupportingClassLoader {
     protected abstract Class<?> loadMockClass(String name, ProtectionDomain protectionDomain);
     
     protected <T> ClassWrapper<T> transformClass(ClassWrapper<T> wrappedType) throws Exception {
-        for (MockTransformer transformer : mockTransformerChain) {
-            wrappedType = transformer.transform(wrappedType);
-        }
+        wrappedType = mockTransformerChain.transform(wrappedType);
+        
         if (classMarker != null) {
             classMarker.mark(wrappedType);
         }
