+++ powermock-core/src/main/java/org/powermock/core/transformers/javassist/StaticFinalNativeMethodMockTransformer.java
@@ -1,0 +1,46 @@
+/*
+ *
+ *   Copyright 2017 the original author or authors.
+ *
+ *   Licensed under the Apache License, Version 2.0 (the "License");
+ *   you may not use this file except in compliance with the License.
+ *   You may obtain a copy of the License at
+ *
+ *       http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *   Unless required by applicable law or agreed to in writing, software
+ *   distributed under the License is distributed on an "AS IS" BASIS,
+ *   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *   See the License for the specific language governing permissions and
+ *   limitations under the License.
+ *
+ */
+
+package org.powermock.core.transformers.javassist;
+
+import javassist.CannotCompileException;
+import javassist.CtClass;
+import javassist.CtMethod;
+import javassist.NotFoundException;
+import org.powermock.core.transformers.TransformStrategy;
+
+import static org.powermock.core.transformers.TransformStrategy.INST_TRANSFORM;
+
+public class StaticFinalNativeMethodMockTransformer extends MethodMockTransformer {
+    
+    public StaticFinalNativeMethodMockTransformer(final TransformStrategy strategy) {
+        super(strategy);
+    }
+    
+    @Override
+    public CtClass transform(final CtClass clazz) throws NotFoundException, CannotCompileException {
+        if (getStrategy() != INST_TRANSFORM) {
+            for (CtMethod m : clazz.getDeclaredMethods()) {
+                modifyMethod(m);
+            }
+        }
+        return clazz;
+    }
+    
+    
+}
