+++ powermock-core/src/main/java/org/powermock/core/transformers/javassist/SuppressStaticInitializerMockTransformer.java
@@ -1,0 +1,45 @@
+/*
+ *
+ *   Copyright 2017 the original author or authors.
+ *
+ *   Licensed under the Apache License, Version 2.0 (the "License");
+ *   you may not use this file except in compliance with the License.
+ *   You may obtain a copy of the License at
+ *
+ *       http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *   Unless required by applicable law or agreed to in writing, software
+ *   distributed under the License is distributed on an "AS IS" BASIS,
+ *   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *   See the License for the specific language governing permissions and
+ *   limitations under the License.
+ *
+ */
+
+package org.powermock.core.transformers.javassist;
+
+import javassist.CannotCompileException;
+import javassist.CtClass;
+import javassist.CtConstructor;
+import org.powermock.core.MockGateway;
+import org.powermock.core.transformers.TransformStrategy;
+
+import static org.powermock.core.transformers.TransformStrategy.CLASSLOADER;
+
+public class SuppressStaticInitializerMockTransformer extends AbstractJavaAssistMockTransformer {
+    
+    public SuppressStaticInitializerMockTransformer(final TransformStrategy strategy) {
+        super(strategy);
+    }
+    
+    @Override
+    public CtClass transform(final CtClass clazz) throws CannotCompileException {
+        if (getStrategy() == CLASSLOADER) {
+            if (MockGateway.staticConstructorCall(clazz.getName()) != MockGateway.PROCEED) {
+                CtConstructor classInitializer = clazz.makeClassInitializer();
+                classInitializer.setBody("{}");
+            }
+        }
+        return clazz;
+    }
+}
