+++ powermock-core/src/test/java/org/powermock/core/transformers/javassist/ClassFinalModifierMockTransformerTest.java
@@ -1,0 +1,59 @@
+/*
+ *
+ *   Copyright 2017 the original author or authors.
+ *
+ *   Licensed under the Apache License, Version 2.0 (the "License");
+ *   you may not use this file except in compliance with the License.
+ *   You may obtain a copy of the License at
+ *
+ *       http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *   Unless required by applicable law or agreed to in writing, software
+ *   distributed under the License is distributed on an "AS IS" BASIS,
+ *   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *   See the License for the specific language governing permissions and
+ *   limitations under the License.
+ *
+ */
+
+package org.powermock.core.transformers.javassist;
+
+import javassist.Modifier;
+import org.junit.Test;
+import org.powermock.core.transformers.TransformStrategy;
+import powermock.test.support.MainMockTransformerTestSupport.SupportClasses;
+
+import static org.junit.Assert.assertFalse;
+
+public class ClassFinalModifierMockTransformerTest extends AbstractBaseMockTransformerTest {
+    
+    @Override
+    protected ClassFinalModifierMockTransformer createMockTransformer() {
+        return new ClassFinalModifierMockTransformer(TransformStrategy.CLASSLOADER);
+    }
+    
+    @Test
+    public void staticFinalInnerClassesShouldBecomeNonFinal() throws Exception {
+        Class<?> clazz = loadWithMockClassLoader(SupportClasses.StaticFinalInnerClass.class.getName());
+        assertFalse(Modifier.isFinal(clazz.getModifiers()));
+    }
+    
+    @Test
+    public void finalInnerClassesShouldBecomeNonFinal() throws Exception {
+        Class<?> clazz = loadWithMockClassLoader(SupportClasses.FinalInnerClass.class.getName());
+        assertFalse(Modifier.isFinal(clazz.getModifiers()));
+    }
+    
+    @Test
+    public void enumClassesShouldBecomeNonFinal() throws Exception {
+        Class<?> clazz = loadWithMockClassLoader(SupportClasses.EnumClass.class.getName());
+        assertFalse(Modifier.isFinal(clazz.getModifiers()));
+    }
+    
+    @Test
+    public void privateInnerClassesShouldBecomeNonFinal() throws Exception {
+        Class<?> clazz = loadWithMockClassLoader(SupportClasses.class.getName() + "$PrivateStaticFinalInnerClass");
+        assertFalse(Modifier.isFinal(clazz.getModifiers()));
+    }
+    
+}
